+++ bpe.h
@@ -2,10 +2,9 @@
 // https://github.com/huggingface/transformers/blob/main/src/transformers/models/biogpt/tokenization_biogpt.py
 #pragma once
 
+#include "biogpt-util.h"
+
 #include <map>
 #include <string>
-
-
-using word_pair = std::pair<std::string, std::string>;
 
 std::string bpe(const std::string& token, std::map<word_pair, int>& bpe_ranks);
